+++ pyfor/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-__version__ = '0.2'
+__version__ = '0.3'
 
 from pyfor import cloud
 from pyfor import rasterizer